+++ client/src/page/User/Login.tsx
@@ -1,17 +1,6 @@
 import React, { useState } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
-import googleIcon from "../../assets/icons/google-icon.svg";
-import naverIcon from "../../assets/icons/naver-icon.svg";
-import kakaoIcon from "../../assets/icons/kakao-icon.svg";
-import {
-	loginContainer,
-	socialLoginButtons,
-	googleButton,
-	kakaoButton,
-	naverButton,
-	iconStyle,
-	joinLink,
-} from "../../page/User/Login.css";
+import { loginContainer, joinLink } from "../../page/User/Login.css";
 import { REGEX } from "./constants/constants";
 import EmailForm from "../../component/User/EmailForm";
 import PasswordForm from "../../component/User/PasswordForm";
@@ -20,9 +9,9 @@ import { useUserStore } from "../../state/store";
 import { ApiCall } from "../../api/api";
 import { ClientError } from "../../api/errors";
 import { sendPostLoginRequest } from "../../api/users/crud";
-import { getOAuthLoginUrl } from "../../api/users/oauth";
 import { ValidateText } from "./Join";
 import ErrorMessageForm from "../../component/User/ErrorMessageForm";
+import OAuthLoginButtons from "../../component/User/OAuthLoginButtons";
 
 const Login: React.FC = () => {
 	const [email, setEmail] = useState<ValidateText>({
@@ -106,45 +95,6 @@ const Login: React.FC = () => {
 		}
 	};
 
-	const handleGoogleLogin = async () => {
-		try {
-			const url = await getOAuthLoginUrl("google");
-			if (url) {
-				window.location.href = url;
-			}
-		} catch (error) {
-			console.error("Google 로그인 에러", error);
-			alert("Google 로그인 실패");
-			navigate("/");
-		}
-	};
-
-	const handleKakaoLogin = async () => {
-		try {
-			const url = await getOAuthLoginUrl("kakao");
-			if (url) {
-				window.location.href = url;
-			}
-		} catch (error) {
-			console.error("kakao 로그인 에러", error);
-			alert("kakao 로그인 실패");
-			navigate("/");
-		}
-	};
-
-	const handleNaverLogin = async () => {
-		try {
-			const url = await getOAuthLoginUrl("naver");
-			if (url) {
-				window.location.href = url;
-			}
-		} catch (error) {
-			console.error("naver 로그인 에러", error);
-			alert("naver 로그인 실패");
-			navigate("/");
-		}
-	};
-
 	return (
 		<div className={loginContainer}>
 			<h1>로그인</h1>
@@ -188,41 +138,7 @@ const Login: React.FC = () => {
       </div> */
 			}
 
-			<div className={socialLoginButtons}>
-				<button
-					className={googleButton}
-					onClick={handleGoogleLogin}
-				>
-					<img
-						src={googleIcon}
-						alt="Google Icon"
-						className={iconStyle}
-					/>
-					구글로 로그인
-				</button>
-				<button
-					className={naverButton}
-					onClick={handleNaverLogin}
-				>
-					<img
-						src={naverIcon}
-						alt="Naver Icon"
-						className={iconStyle}
-					/>
-					네이버로 로그인
-				</button>
-				<button
-					className={kakaoButton}
-					onClick={handleKakaoLogin}
-				>
-					<img
-						src={kakaoIcon}
-						alt="Kakao Icon"
-						className={iconStyle}
-					/>
-					카카오로 로그인
-				</button>
-			</div>
+			<OAuthLoginButtons loginType="login" />
 		</div>
 	);
 };
